Adding selectors folder
